requirements.txt
appnope==0.1.0
astroid==2.3.3
attrs==19.3.0
audio-metadata==0.10.0
audioread==2.1.8
autopep8==1.5.2
backcall==0.1.0
bidict==0.19.0
bitstruct==8.10.0
bleach==3.1.4
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
eyeD3==0.8.12
ffmpeg==1.4
Flask==1.1.2
Flask-Cors==3.0.8
hmmlearn==0.2.2
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
librosa==0.7.2
llvmlite==0.31.0
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
numba==0.48.0
numpy==1.18.2
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pendulum==2.0.4
pexpect==4.8.0
pickleshare==0.7.5
plac==1.1.3
pprintpp==0.4.0
praat-parselmouth==0.3.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.20
pydub==0.23.1
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-magic==0.4.15
pytz==2019.3
pytzdata==2019.3
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
resampy==0.2.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
sentencepiece==0.1.85
simplejson==3.16.0
six==1.14.0
SoundFile==0.10.3.post1
spacy==2.2.4
srsly==1.0.2
tbm-utils==2.5.1
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
torch==1.4.0
torchtext==0.5.0
tornado==6.0.4
tqdm==4.44.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wit==5.1.0
wrapt==1.11.2
zipp==3.1.0
